[#25] Refactor: BookRawPublishPort common으로 위치 변경 및 그에 따른 import 수정
